feat(settings): left-rail IA reorganisation (#215)
Collapses the 22-tab horizontal scroll strip at /settings into a 6-group
left-rail layout with nested routes, per-panel code splitting, URL deep
linking, and `aria-current="page"` a11y. Also adds backward-compat
redirects for legacy `?tab=<id>` URLs.

Categories (6 groups, 22 items — 1:1 panel mapping, nothing removed):
- Personal (3): Confluence · AI Prompts · Theme
- Content (3): Spaces · Sync · Labels
- AI (6): LLM · Embedding · AI Safety · Workers · LLM Policy (EE) · LLM Audit (EE)
- Integrations (3): Email/SMTP · SearXNG · MCP Docs
- Security & Access (4): Rate Limits · SSO (EE) · SCIM (EE) · Data Retention (EE)
- System (3): License · Errors · System

URL structure: `/settings/<category>/<item>` is now the source of truth.
Legacy `?tab=<id>` redirects client-side via a useSearchParams effect in
SettingsLayout using the `legacyTabMap` generated from the nav config.
The `ollama` tab id → `llm` (provider-agnostic now); handled by the map.

Architecture:
- `settings-nav.ts` — single source of truth (grouped nav + legacy map
  + `canSeeItem` + `firstVisiblePath`).
- `SettingsLayout.tsx` — <nav aria-label="Settings"> landmark + <Outlet/>
  per W3C APG guidance. Group headers are real <h2>. Keyboard nav is
  browser-native Tab (no roving tabindex needed for link navigation).
  `motion-safe:` Tailwind variants respect prefers-reduced-motion
  (ADR-010).
- `SettingsIndexRedirect` — bounces `/settings` to the first visible
  panel the user has permission for.
- `SettingsPanelRoute.tsx` — resolves `:category/:item` to a lazy()
  panel via a registry, guards gating (admin/EE/feature-flag) against
  direct-URL access.
- `panels/` — 6 inline panels (ConfluenceTab, SyncTab, LlmTab,
  EmbeddingTab, AiPromptsTab, SystemTab) extracted from the monolithic
  SettingsPage.tsx so each gets its own route-level chunk. The rest
  already lived in their own files.
- 22 separate JS chunks confirmed in `npm run build` output.

Accessibility:
- `<nav>` landmark + group headings + `aria-current="page"` on the
  active link. No ARIA tabs pattern — this is URL-driven navigation,
  not in-page view switching (per APG).
- EE-only items get a visible "EE" badge.
- Reduced-motion respected via motion-safe: throughout.

Backward compat:
- `?tab=<id>` redirects (exhaustive table in plan §3).
- Direct URL access to admin/EE-only panels by a non-eligible user
  bounces to the default; unknown paths fall through the same way.
- `SettingsPage.tsx` is kept as a 180-line test harness for 4 legacy
  tests (SettingsPage.test.tsx, SyncTab.test.tsx, ThemeTab.test.tsx,
  OllamaTab.test.tsx) that mount it directly. To be removed in a
  follow-up once those tests migrate to SettingsLayout.
- `usecase-*` data-testids from PR #216 (issue #214) are preserved
  byte-identical.
- Rail links carry both `data-testid="nav-settings-<id>"` and
  `data-legacy-testid="tab-<id>"` for one release (removal tracked
  in plan §15.1).

Planning + research:
- docs/plans/215-settings-ia-reorg.md (392 lines) — the full plan.
- Research synthesised from Explore agent (codebase inventory),
  Gemini (GitLab/Grafana/Linear/Jira/GitHub/Notion IA teardown,
  NNGroup / Polaris / Primer heuristics), and docs-researcher via
  Ref MCP (Radix Tabs, React Router v7 nested routes + redirect,
  Tailwind 4 `motion-safe:`, W3C APG Tabs/Navigation).

Tests:
- `SettingsLayout.test.tsx` (new) — 11 tests covering rail visibility,
  admin/EE gating, `aria-current`, index redirect, legacy ?tab=
  redirect, direct-URL gating.
- `e2e/settings-nav.spec.ts` (new) — deep link, keyboard nav, legacy
  redirect, back/forward, admin group visibility, direct-URL bounce.
- All 112 existing settings unit tests untouched and passing.
- Frontend total: 1794/1794 pass (up from 1783).
- `npm run typecheck` + `npm run lint` + `npm run build` all clean.

Out of scope (per issue #215):
- Backend: zero changes.
- Panel internals: untouched; reshuffle only.
