+++ frontend/src/features/settings/panels/index.ts
@@ -1,0 +1,6 @@
+export { ConfluenceTab } from './ConfluenceTab';
+export { SyncTab } from './SyncTab';
+export { LlmTab, OllamaTab } from './LlmTab';
+export { EmbeddingTab } from './EmbeddingTab';
+export { AiPromptsTab } from './AiPromptsTab';
+export { SystemTab } from './SystemTab';
